# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HouseHunt - House Rent Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  owner: 'owner',
  renter: 'renter',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  isApproved: boolean;
  createdAt: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  owner: 'owner',
  renter: 'renter',
} as const;

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export const PropertyType = {
  apartment: 'apartment',
  house: 'house',
  villa: 'villa',
  studio: 'studio',
  commercial: 'commercial',
} as const;

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


export const PropertyStatus = {
  available: 'available',
  booked: 'booked',
} as const;

export interface Property {
  id: number;
  title: string;
  description: string;
  location: string;
  rent: number;
  type: PropertyType;
  amenities?: string[];
  images?: string[];
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  status: PropertyStatus;
  createdAt: string;
}

export type PropertyInputType = typeof PropertyInputType[keyof typeof PropertyInputType];


export const PropertyInputType = {
  apartment: 'apartment',
  house: 'house',
  villa: 'villa',
  studio: 'studio',
  commercial: 'commercial',
} as const;

export interface PropertyInput {
  /** @minLength 3 */
  title: string;
  description: string;
  location: string;
  rent: number;
  type: PropertyInputType;
  amenities?: string[];
  images?: string[];
}

export type PropertyUpdateType = typeof PropertyUpdateType[keyof typeof PropertyUpdateType];


export const PropertyUpdateType = {
  apartment: 'apartment',
  house: 'house',
  villa: 'villa',
  studio: 'studio',
  commercial: 'commercial',
} as const;

export type PropertyUpdateStatus = typeof PropertyUpdateStatus[keyof typeof PropertyUpdateStatus];


export const PropertyUpdateStatus = {
  available: 'available',
  booked: 'booked',
} as const;

export interface PropertyUpdate {
  title?: string;
  description?: string;
  location?: string;
  rent?: number;
  type?: PropertyUpdateType;
  amenities?: string[];
  images?: string[];
  status?: PropertyUpdateStatus;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  propertyId: number;
  renterId: number;
  startDate: string;
  endDate: string;
  status: BookingStatus;
  /** @nullable */
  propertyTitle?: string | null;
  /** @nullable */
  propertyLocation?: string | null;
  /** @nullable */
  propertyRent?: number | null;
  /** @nullable */
  renterName?: string | null;
  /** @nullable */
  renterEmail?: string | null;
  createdAt: string;
}

export interface BookingInput {
  propertyId: number;
  startDate: string;
  endDate: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface ApprovalUpdate {
  isApproved: boolean;
}

export interface AdminStats {
  totalUsers: number;
  totalOwners: number;
  totalRenters: number;
  totalProperties: number;
  totalBookings: number;
  pendingBookings: number;
  availableProperties: number;
  bookedProperties: number;
}

export interface OwnerStats {
  totalProperties: number;
  availableProperties: number;
  bookedProperties: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
}

export type ListPropertiesParams = {
location?: string;
minRent?: number;
maxRent?: number;
type?: string;
status?: string;
};