# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HouseHunt - House Rent Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or email taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a property (renter only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid dates or property unavailable
        "401":
          description: Unauthorized
  /bookings/my:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: Get bookings for the current renter
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Renter's bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/owner:
    get:
      operationId: listOwnerBookings
      tags: [bookings]
      summary: Get bookings for properties owned by current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Owner's bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (owner or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/approve:
    patch:
      operationId: adminApproveOwner
      tags: [admin]
      summary: Approve or reject an owner (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /admin/properties:
    get:
      operationId: adminListProperties
      tags: [admin]
      summary: List all properties (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin]
      summary: List all bookings (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get platform statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isApproved, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, owner, renter]
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [owner, renter]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Property:
      type: object
      required: [id, title, description, location, rent, type, status, ownerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        rent:
          type: number
        type:
          type: string
          enum: [apartment, house, villa, studio, commercial]
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, booked]
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      required: [id, propertyId, renterId, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        renterId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        propertyTitle:
          type: ["string", "null"]
        propertyLocation:
          type: ["string", "null"]
        propertyRent:
          type: ["number", "null"]
        renterName:
          type: ["string", "null"]
        renterEmail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [propertyId, startDate, endDate]
      properties:
        propertyId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
    ApprovalUpdate:
      type: object
      required: [isApproved]
      properties:
        isApproved:
          type: boolean
    AdminStats:
      type: object
      required: [totalUsers, totalOwners, totalRenters, totalProperties, totalBookings, pendingBookings, availableProperties, bookedProperties]
      properties:
        totalUsers:
          type: integer
        totalOwners:
          type: integer
        totalRenters:
          type: integer
        totalProperties:
          type: integer
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        availableProperties:
          type: integer
        bookedProperties:
          type: integer
...