# lib/api-zod/src/generated/types/approvalUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HouseHunt - House Rent Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ApprovalUpdate {
  isApproved: boolean;
}