# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HouseHunt - House Rent Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  propertyId: number;
  renterId: number;
  startDate: Date;
  endDate: Date;
  status: BookingStatus;
  /** @nullable */
  propertyTitle?: string | null;
  /** @nullable */
  propertyLocation?: string | null;
  /** @nullable */
  propertyRent?: number | null;
  /** @nullable */
  renterName?: string | null;
  /** @nullable */
  renterEmail?: string | null;
  createdAt: Date;
}