# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HouseHunt - House Rent Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from './propertyStatus';
import type { PropertyType } from './propertyType';

export interface Property {
  id: number;
  title: string;
  description: string;
  location: string;
  rent: number;
  type: PropertyType;
  amenities?: string[];
  images?: string[];
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  status: PropertyStatus;
  createdAt: Date;
}